requirements.txt
selenium==4.25.0
PyYAML==6.0.2
